requirements.txt
#Librerias para el funcionamineto del backend
fastapi
uvicorn 
python-dotenv
pydantic
httpx

#Librerias para el manejo de archivos
python-docx
docxtpl
docxcompose
pillow

PyPDF2
PyMuPDF
pypandoc
python-multipart # lee y decodifica los archivos enviados


diagrams
#npm install -g @mermaid-js/mermaid-cli

